refactor: move symbol parsing out of request handler
